0049[M]: Add hash solution
Use a hash table and use the sorted string as the hash key, and anagram
strings must have the same value. Then traverse all strings and update the hash
table, also recording the original string for the purpose of the final result.

Finally, traverse the available hash buckets and create the final result.