0930[M]: Add two O(n) solutions
Hash:
Using a hash to map the occurrence of each cumulative sum, initialize
subarr_sum_set[0] = 1 in this mapping pair. While traversing the nums
array, for each encountered number num, add it to the cumulative sum
tmp_sum. Then, check if there is a mapping value for tmp_sum - goal. If
so, it implies that there are subarr_sum_set[tmp_sum-goal] subarrays that
meet the requirements, which should be added to the result ret. If tmp_sum
equals goal, i.e., tmp_sum - goal = 0, it indicates that from the beginning
to the current position is exactly the required subarray.

Sliding Window:
Utilizes the concept of prefix sum. However, this sum isn't the sum from
the beginning to the current position, but rather the sum within the
sliding window. Whenever the sum within this sliding window equals the
goal, it constitutes a valid subarray according to the problem.

Traverse the nums array, adding the current digit to the sum. If the sum
exceeds the goal at any point, it requires shrinking the window by moving
the left to the right and subtracting the digit removed from the window until
the sum is less than or equal to goal, if the sum is less than the goal,
it implies that there is no subarray whose sum equals the goal. If the
sum equals the goal, increment the result (ret) by 1. Another scenario
to consider is when there are consecutive zeros on the left side of the
window. Zeros don't affect the sum, but each zero contributes to a
distinct subarray. Hence, it's necessary to count the number of consecutive
zeros from the left.