1255[H]: Add backtracking solution
Approach: Backtrack

The recursive function takes four parameters: the original word list words, a
count of available letters cnts, the letter score array score, and the current
index idx. Starting from idx, we iterate through the words. For each word, we
check each letter and decrement the corresponding count in cnts. If a count goes
below zero, it means we can't form the current word, and we set is_valid to
false. We also calculate the score for the current word and store it in sum. If
is_valid remains true after processing the word, we recursively call the function
for the next index and add the result to sum. This valid sum is used to update
the result res. Finally, we restore the counts by adding back the letters of the
current word.