0140[H]: Add recursive DP solution
Solution Approach:

Use a structure str_arr_t to store the result sentences. This struct includes
an array of strings (data), the current size of the array (size), and its
capacity (capacity). Initialize the str_arr_t with a specific capacity and
provide functions to add strings to this array. Helper Functions:

The main recursive function wb is designed to explore all possible ways to
break the string word into valid words found in the dictionary. The function
operates by iterating over all possible prefixes of the current string. For
each prefix that is found in the dictionary, it calls itself recursively with
the remaining part of the string. During each recursive call, it constructs a
new sentence by appending the valid prefix to the current result.

When the length of the current string(word) is zero, it means a valid sentence
has been formed using words from the dictionary. This sentence is added to the
result array.