1209[M]: add simple stack solution
Approach: Stack

We define a structure data_t to store characters (ch) and their consecutive
occurrence counts (consec_occ). Then, we initialize a stack (stk) to store
characters and their consecutive occurrence counts. We iterate through each
character in the input string s, pushing the current character onto the stack
with a consecutive occurrence count of 1 if the stack is empty or if the current
character is different from the top character of the stack. If the current
character is the same as the top character of the stack, we increment its
consecutive occurrence count unless it equals k - 1, in which case we pop it
from the stack.