0502[H]: Add heap solution
The solution utilizes a combination of min-heap and max-heap to efficiently
manage and select projects based on their capital requirements and profits.
Initially, all projects are inserted into a min-heap, which sorts them by their
capital requirements. This allows us to quickly find the projects that can be
started with the current available capital. In each iteration, projects that can
be started (those with capital requirements less than or equal to the current
available capital) are moved from the min-heap to the max-heap. The max-heap,
which sorts projects by their profit, ensures that we always select the project
with the highest profit that can be started with the available capital. This
project is then executed, and its profit is added to the available capital. This
process is repeated up to k times or until no more projects can be started. The
final available capital is returned as the result. This approach ensures that we
maximize the capital by always choosing the most profitable projects that can be
started with the current capital.