1438[M]: Add sliding windown and monotonic queue solution
Approach: Sliding Window + Monotonic Queue

Maintain a monotonically decreasing queue, maxq, and a monotonically increasing
queue, minq, where both store indices. The front element of maxq is the index of
the current maximum element, and the front element of minq is the index of the
current minimum element. Note that while storing elements is possible, storing
indices is beneficial because it remains unaffected by duplicate elements.

Also, need two pointers, left and i. Initially, i moves forward. When it finds
that:
- maxq is not empty and the last element in maxq is smaller than the current
  element nums[i], it continually polls elements from maxq until it becomes
  monotonically decreasing again.
- minq is not empty and the last element in minq is larger than the current
  element nums[i], it continually polls elements from minq until it becomes
  monotonically increasing again.

At this point, we check: if both queues are not empty, but the diff between
the front elements in the two queues (one being the maximum value and the other
the minimum value) is greater than the limit, we start moving the left pointer
to the right. As we move the left pointer, if any element in the queues has an
index less than or equal to left, we poll it out because it's no longer needed.
Here, storing indices is advantageous because the indices are sequentially
added to both queues, so if the diff exceeds the limit, we need to start
checking from the very beginning of the left pointer.

Time complexity: O(n)
Space complexity: O(n)