2192[M]: Add DFS solution
Approach: DFS

Solved this problem using a method similar to reverse DFS. The task is to find
the ancestors of each node in the graph. Since the graph itself is a Directed
Acyclic Graph (DAG), and each edge has a direction, we can reverse the direction
of each edge when building the graph to find the ancestors. For example, if
there is an edge from A to B, we can add A to the adjacency list of B when
constructing the graph. This means that if we start from B, we can reach its
ancestor A.